abstract medicine box endpoints
